langchain-gemini-project/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
faiss-cpu
chromadb
python-dotenv
pypdf
